finnhub-producer/requirements.txt
websocket-client
confluent-kafka
finnhub-python